gym_agent_uv/pyproject.toml
[project]
name = "gym-agent"
version = "0.1.0"
description = "Find gym / fitness / pilates leads from Instagram + Google → Google Sheets"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-auth>=2.49.1",
    "google-search-results>=2.4.2",
    "gspread>=6.2.1",
    "httpx>=0.28.1",
    "instaloader>=4.15.1",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
    "webdriver-manager>=4.0.2",
]

[project.scripts]
gym-agent = "gym_agent.main:cli"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gym_agent"]

[dependency-groups]
dev = [
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
